# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MCP Super-Server API
  description: |
    Voice-native Model Context Protocol Super-Server with Multimodal Context Fabric,
    Multi-Agent Orchestration, Event-Sourced World State, and Cross-Platform Identity Mesh.
  version: 0.0.1
  contact:
    name: MCP Super-Server Team
  license:
    name: Proprietary
    url: https://example.com/license
servers:
  - url: http://localhost:3099
    description: Local development
...
paths:
...
  /tools/invoke:
    post:
      tags: [Tools]
      summary: Invoke a tool through the policy gate
      operationId: invokeTool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionId, toolId, input]
              properties:
                sessionId:
                  type: string
                toolId:
                  type: string
                  description: Tool identifier (e.g., 'web:fetch', 'db:query')
                input:
                  type: object
                  description: Tool-specific input parameters
      responses:
        '200':
          description: Tool invocation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResult'
...
components:
  schemas:
...
    ToolResult:
      type: object
      properties:
        decision:
          type: string
          enum: [allow, deny, require_human]
        result:
          type: object
          nullable: true
        reason:
          type: string
          nullable: true
...